requirements.txt
elastalert>=0.1.17
kombu==4.1.0
pytz>=2017.2
python-dateutil==2.6.1
